requirements.txt
EMD_signal==1.2.3
EntropyHub==0.2
hmmlearn==0.3.3
keras==2.10.0
matplotlib==3.9.2
minepy==1.2.6
numpy==1.26.4
pandas==2.2.2
platypus==1.0
pydantic==2.10.5
pyDOE==0.3.8
scikit_learn==1.3.0
scipy==1.14.1
seaborn==0.13.2
statsmodels==0.13.5
tensorflow-gpu==2.10.0
typing_extensions==4.12.2
vmdpy==0.2
xgboost==2.0.0
